# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/internal/admin/organisations/{organisationId}/suspend" : {
      "post" : {
        "tags" : [ "admin-organisation-controller" ],
        "operationId" : "suspend",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganisationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/organisations/{organisationId}/reinstate" : {
      "post" : {
        "tags" : [ "admin-organisation-controller" ],
        "operationId" : "reinstate",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganisationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{organisationId}" : {
      "get" : {
        "tags" : [ "organisation-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganisationView"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "organisation-controller" ],
        "operationId" : "rename",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Name"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganisationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{organisationId}/members/{userId}" : {
      "delete" : {
        "tags" : [ "organisation-controller" ],
        "operationId" : "remove",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "organisation-controller" ],
        "operationId" : "role",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangeRole"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MemberView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{organisationId}/members" : {
      "get" : {
        "tags" : [ "organisation-controller" ],
        "operationId" : "members",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SliceMemberView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/organisations/{organisationId}/members" : {
      "get" : {
        "tags" : [ "internal-organisation-controller" ],
        "operationId" : "members_1",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "permission",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "ORG_VIEW", "ORG_MANAGE", "MEMBERS_MANAGE", "BUSINESS_APPLY", "OUTLET_MANAGE", "MENU_MANAGE", "STOCK_TOGGLE", "ORDERS_OPERATE", "EARNINGS_VIEW", "PAYOUTS_MANAGE", "ADS_VIEW", "ADS_MANAGE", "WALLET_VIEW", "WALLET_TOPUP" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string",
                    "format" : "uuid"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/organisations/{organisationId}/members/{userId}" : {
      "get" : {
        "tags" : [ "internal-organisation-controller" ],
        "operationId" : "membership",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MembershipDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/organisations" : {
      "get" : {
        "tags" : [ "admin-organisation-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SliceOrganisationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/organisations/{organisationId}" : {
      "get" : {
        "tags" : [ "admin-organisation-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminOrganisationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/organisations/{organisationId}/members" : {
      "get" : {
        "tags" : [ "admin-organisation-controller" ],
        "operationId" : "members_2",
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SliceMemberView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/audit-events" : {
      "get" : {
        "tags" : [ "admin-audit-controller" ],
        "operationId" : "history",
        "parameters" : [ {
          "name" : "subjectType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SliceAuditEvent"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Name" : {
        "required" : [ "displayName" ],
        "type" : "object",
        "properties" : {
          "displayName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "OrganisationView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "displayName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "SUSPENDED", "CLOSED" ]
          },
          "myRole" : {
            "type" : "string",
            "enum" : [ "OWNER", "ADMIN", "MANAGER", "STAFF" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Reason" : {
        "required" : [ "reason" ],
        "type" : "object",
        "properties" : {
          "reason" : {
            "maxLength" : 500,
            "minLength" : 10,
            "type" : "string"
          }
        }
      },
      "ChangeRole" : {
        "required" : [ "role" ],
        "type" : "object",
        "properties" : {
          "role" : {
            "type" : "string",
            "enum" : [ "OWNER", "ADMIN", "MANAGER", "STAFF" ]
          }
        }
      },
      "MemberView" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "OWNER", "ADMIN", "MANAGER", "STAFF" ]
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "REMOVED" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        }
      },
      "SliceOrganisationView" : {
        "type" : "object",
        "properties" : {
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrganisationView"
            }
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
      "SliceMemberView" : {
        "type" : "object",
        "properties" : {
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MemberView"
            }
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
      "MembershipDto" : {
        "required" : [ "organisationId", "organisationStatus", "role", "status", "userId" ],
        "type" : "object",
        "properties" : {
          "organisationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "organisationStatus" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "SUSPENDED", "CLOSED" ]
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "OWNER", "ADMIN", "MANAGER", "STAFF" ]
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "REMOVED" ]
          }
        }
      },
...
      "AdminOrganisationView" : {
        "type" : "object",
        "properties" : {
          "organisation" : {
            "$ref" : "#/components/schemas/OrganisationView"
          },
          "members" : {
            "$ref" : "#/components/schemas/SliceMemberView"
          }
        }
      },
      "AuditEvent" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "actorUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "actorKind" : {
            "type" : "string"
          },
          "action" : {
            "type" : "string",
            "enum" : [ "ORGANISATION_CREATED", "ORGANISATION_RENAMED", "ORGANISATION_SUSPENDED", "ORGANISATION_REINSTATED", "ORGANISATION_CLOSED", "MEMBER_INVITED", "INVITATION_ACCEPTED", "INVITATION_DECLINED", "INVITATION_REVOKED", "INVITATION_EXPIRED", "MEMBER_ROLE_CHANGED", "MEMBER_REMOVED", "OWNERSHIP_TRANSFERRED", "APPLICATION_SUBMITTED", "APPLICATION_APPROVED", "APPLICATION_REJECTED", "APPLICATION_SUSPENDED", "APPLICATION_REINSTATED", "KYC_DOCUMENT_VIEWED", "ADMIN_STEP_UP", "TOPUP_REQUESTED", "TOPUP_SETTLED", "TOPUP_FAILED", "CREATIVE_APPROVED", "CREATIVE_REJECTED", "CAMPAIGN_ACTIVATED", "CAMPAIGN_PAUSED" ]
          },
          "subjectType" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "organisationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reason" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        }
      },
      "SliceAuditEvent" : {
        "type" : "object",
        "properties" : {
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditEvent"
            }
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}